refact : BallsTest, GameTest에 있는 TestCase 중복이기에 제거
