+++ src/test/java/baseball/domain/BallsTest.java
@@ -1,14 +1,10 @@
 package baseball.domain;
 
+import baseball.domain.fixture.MatchingBalls;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -48,95 +44,13 @@ public class BallsTest {
     }
 
     @ParameterizedTest
-    @MethodSource("createMatchingPackageSource")
+    @MethodSource("baseball.domain.fixture.MatchingBalls#createMatchingPackageSource")
     @DisplayName("공들과 공들은 매칭하여, 결과를 도출 할 수 있다.")
-    public void 공들과_공들은_매칭하여_결과를_도출_할_수_있다(MatchingPackage matchingPackage) {
+    public void 공들과_공들은_매칭하여_결과를_도출_할_수_있다(MatchingBalls.MatchingPackage matchingPackage) {
         Score match = matchingPackage.getSource().match(matchingPackage.getDest());
 
         assertThat(match.countOfStrike()).isEqualTo(matchingPackage.getCountOfStrike());
         assertThat(match.countOfBall()).isEqualTo(matchingPackage.getCountOfBall());
         assertThat(match.countOfMissing()).isEqualTo(matchingPackage.getCountOfMissing());
-    }
-
-    private static List<MatchingPackage> createMatchingPackageSource() {
-        List<MatchingPackage> matchingPackages = new ArrayList<>();
-
-        matchingPackages.add(new MatchingPackage(
-                Arrays.asList(7, 1, 3),
-                Arrays.asList(1, 2, 3),
-                1,
-                1,
-                1
-        ));
-
-        matchingPackages.add(new MatchingPackage(
-                Arrays.asList(7, 1, 3),
-                Arrays.asList(1, 4, 5),
-                0,
-                1,
-                2
-        ));
-
-        matchingPackages.add(new MatchingPackage(
-                Arrays.asList(7, 1, 3),
-                Arrays.asList(6, 7, 1),
-                0,
-                2,
-                1
-        ));
-
-        matchingPackages.add(new MatchingPackage(
-                Arrays.asList(7, 1, 3),
-                Arrays.asList(2, 1, 6),
-                1,
-                0,
-                2
-        ));
-
-        matchingPackages.add(new MatchingPackage(
-                Arrays.asList(7, 1, 3),
-                Arrays.asList(7, 1, 3),
-                3,
-                0,
-                0
-        ));
-
-        return matchingPackages;
-    }
-
-    private static class MatchingPackage {
-        private Balls source;
-        private Balls dest;
-        private int countOfStrike;
-        private int countOfBall;
-        private int countOfMissing;
-
-        public MatchingPackage(List<Integer> source, List<Integer> dest, int countOfStrike, int countOfBall, int countOfMissing) {
-            this.source = new Balls(source.stream().map(Ball::of).collect(Collectors.toList()));
-            this.dest = new Balls(dest.stream().map(Ball::of).collect(Collectors.toList()));
-            this.countOfStrike = countOfStrike;
-            this.countOfBall = countOfBall;
-            this.countOfMissing = countOfMissing;
-        }
-
-        public Balls getSource() {
-            return source;
-        }
-
-        public Balls getDest() {
-            return dest;
-        }
-
-        public int getCountOfStrike() {
-            return countOfStrike;
-        }
-
-        public int getCountOfBall() {
-            return countOfBall;
-        }
-
-        public int getCountOfMissing() {
-            return countOfMissing;
-        }
     }
 }
